pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crud-task"
version = "0.1.0"
description = "Minimal FastAPI CRUD project"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "python-jose>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "psycopg[binary]>=3.2.0",
    "asyncpg>=0.29.0",
    "aioboto3>=13.0.0",
    "uvicorn>=0.30.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "aiosqlite>=0.20.0",
]

[tool.setuptools]
packages = ["app"]
